Created component for slides
