+++ frontend/src/app/components/common/attachments-carousel.tsx
@@ -2,28 +2,23 @@ import { FC, useEffect, useState, useRef } from "react";
 import {
   CarouselProvider,
   Slider,
-  Slide,
   ButtonBack,
   ButtonNext,
 } from "pure-react-carousel";
 import "pure-react-carousel/dist/react-carousel.es.css";
-import { useAppDispatch } from "../../../hooks/hooks";
+import { useAppDispatch } from "../../hooks/hooks";
 import { useSelector } from "react-redux";
-import { RootState } from "../../../store/store";
+import { RootState } from "../../store/store";
 import {
   getAttachments,
   setAttachments,
-} from "../../../store/reducers/objectstore";
-import { generateApiParameters, get } from "../../../common/api";
-import { formatDateTime } from "../../../common/methods";
+} from "../../store/reducers/objectstore";
 import {
-  BsImageFill,
-  BsCloudDownload,
-  BsTrash,
   BsArrowLeftShort,
   BsArrowRightShort,
   BsPlus,
 } from "react-icons/bs";
+import { AttachmentSlide } from "./attachment-slide";
 
 type Props = {
   complaintIdentifier: string;
@@ -42,20 +37,22 @@ export const AttachmentsCarousel: FC<Props> = ({
   );
 
   const SLIDE_WIDTH = 289; // width of the carousel slide, in pixels
+  const [visibleSlides, setVisibleSlides] = useState<number>(4); // Adjust the initial number of visible slides as needed
+  const carouselContainerRef = useRef<HTMLDivElement | null>(null); // ref to the carousel's container, used to determine how many slides can fit in the container
 
+
+  // get the attachments when the complaint loads
   useEffect(() => {
     dispatch(getAttachments(complaintIdentifier));
   }, [complaintIdentifier, dispatch]);
 
+  //-- when the component unmounts clear the complaint from redux
   useEffect(() => {
-    //-- when the component unmounts clear the complaint from redux
     return () => {
       dispatch(setAttachments({}));
     };
   }, []);
 
-  const [visibleSlides, setVisibleSlides] = useState<number>(4); // Adjust the initial number of visible slides as needed
-  const carouselContainerRef = useRef<HTMLDivElement | null>(null);
 
   useEffect(() => {
     // Function to update the number of visible slides based on the parent container width
@@ -98,22 +95,6 @@ export const AttachmentsCarousel: FC<Props> = ({
     };
   }, []);
 
-  const handleImageClick = async (objectid: string, filename: string) => {
-    const parameters = generateApiParameters(
-      `${process.env.REACT_APP_COMS_URL}/object/${objectid}?download=url`
-    );
-    console.log(`${process.env.REACT_APP_COMS_URL}/object`);
-    const response = await get<string>(dispatch, parameters);
-
-    // Create an anchor element to trigger the download.  Note that the href is a pre-signed URL valid for 7 days (this is a COMS/Objectstore feature)
-    const a = document.createElement("a");
-
-    a.href = response;
-    a.download = filename; // Set the download filename
-    a.target = "_blank";
-    a.click();
-  };
-
   return (
     <div className="comp-complaint-details-block" ref={carouselContainerRef}>
       <h6>Attachments ({carouselData?.length ? carouselData.length : 0})</h6>
@@ -141,31 +122,8 @@ export const AttachmentsCarousel: FC<Props> = ({
           )}
           <Slider className="coms-slider">
             {carouselData?.map((item, index) => (
-              <Slide index={index} key={index}>
-                <div className="coms-carousel-slide">
-                  <div className="coms-carousel-actions">
-                    {allowDelete && (
-                      <BsTrash className="delete-icon" tabIndex={index} />
-                    )}
-                    <BsCloudDownload
-                      tabIndex={index}
-                      className="download-icon"
-                      onClick={() =>
-                        handleImageClick(`${item.id}`, `${item.name}`)
-                      }
-                    />
-                  </div>
-                  <div className="top-section">
-                    <BsImageFill />
-                  </div>
-                  <div className="bottom-section">
-                    <div className="line bold">{item.name}</div>
-                    <div className="line">
-                      {formatDateTime(item.createdAt.toString())}
-                    </div>
-                  </div>
-                </div>
-              </Slide>
+              <AttachmentSlide attachment={item} index={index} allowDelete={allowDelete}/>
+
             ))}
           </Slider>
         </CarouselProvider>
